企業が未登録の場合のES作成機能の挙動を修正
